+++ bin/node/src/commands/block_producer.rs
@@ -30,7 +30,7 @@ pub enum BlockProducerCommand {
         /// This can be further configured using environment variables as defined in the official
         /// OpenTelemetry documentation. See our operator manual for further details.
         #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
-        open_telemetry: bool,
+        enable_otel: bool,
     },
 }
 
@@ -41,7 +41,7 @@ impl BlockProducerCommand {
             store_url,
             ntx_builder_url,
             block_producer,
-            open_telemetry: _,
+            enable_otel: _,
         } = self;
 
         let store_address = store_url
@@ -73,7 +73,7 @@ impl BlockProducerCommand {
     }
 
     pub fn is_open_telemetry_enabled(&self) -> bool {
-        let Self::Start { open_telemetry, .. } = self;
-        *open_telemetry
+        let Self::Start { enable_otel, .. } = self;
+        *enable_otel
     }
 }