+++ bin/node/src/commands/bundled.rs
@@ -54,7 +54,7 @@ pub enum BundledCommand {
         /// This can be further configured using environment variables as defined in the official
         /// OpenTelemetry documentation. See our operator manual for further details.
         #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
-        open_telemetry: bool,
+        enable_otel: bool,
     },
 }
 
@@ -76,7 +76,7 @@ impl BundledCommand {
                 data_directory,
                 block_producer,
                 ntx_builder,
-                open_telemetry: _,
+                enable_otel: _,
             } => Self::start(rpc_url, data_directory, ntx_builder, block_producer).await,
         }
     }
@@ -213,8 +213,8 @@ impl BundledCommand {
     }
 
     pub fn is_open_telemetry_enabled(&self) -> bool {
-        if let Self::Start { open_telemetry, .. } = self {
-            *open_telemetry
+        if let Self::Start { enable_otel, .. } = self {
+            *enable_otel
         } else {
             false
         }