+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [BREAKING] Remove default value account filepath for faucet commands and rename --output-path to --output (#976).
 - Exposed `miden-proving-service` as a library (#956).
 - Removed system monitor from node binary (#1019).
+- [BREAKING] Renamed `open_telemetry` to `enable_otel` in all node's commands (#1019).
 
 ### Fixes